Reset skip-to-content link when JS disabled
The existing skip-to-content link jumps to an element that is hidden when JS does not load. Created a new skip-to-content button that jumps to the first element in the pattern category menu when JS disabled. Added new ID to first item in pattern category menu list (id will be there always, not just when no script). Added styling for new skip-to-content button.