Changed method of accessing alt text from pattern md files
Purpose of changes on this branch is to refactor the pattern page to make code which renders the image easier to read. In this commit, changed method of writing and reading alt text for images from the pattern markdown page. Moved alt text to its own label on md page, moving it out of image array. Reduced all switch cases and for loop limits on pattern html template. Replaced old references to alt text with new alt text liquid variable.