Updated 320px media query
Had initially implemented 320px media query for accessibility reasons (to ensure page was still useable if user zoomed in to an equivalent width of 320px). This caused knock-on layout problems on mobile devices at 320px width (iPhone 5). Earlier fix was to knock down media breakpoint to 319px, so that styles still applied on zoom, but were not being applied on phones at 320px width. This was a pretty hack-y solution, so this commit has reset the 320px media query for 320px-width phones. Functionality zooming in to 320px is slightly reduced; will try to work on this in subsequent changes.