Move Vimeo API Script to About page
Vimeo API script tag was included in the HTML head template. As such, it was being loaded on every page even though it was only required on the About page. Early performance tests indicate that render-blocking resources were causing the page load time slow considerably. Put the script tag at the bottom of the About page body and included 'async' attribute as the video is 'below the fold' and doesn't needed to be loaded immediately.