Updated JS and SCSS files to enable jQuery UI removal
jquery-ui.js was one of the render-blocking JS files highlighted in performance testing. The only part of the current app.js file
using it was the mobile menu open/close functions that used the .animate() method to make the header background-color white when
the menu was open, and transparent when the menu was closed. Handled this by adding classes in app.scss to handle background-color
transition, and toggling classes with jQuery in app.js. Only change is transition is not handled quite as smoothly as transition
(would have to be subjective decision on whether or not to integrate these changes into main code).