Updated category names and descriptions; updated homepage layout to v2
Updated the category names to their new versions, also added descriptions for each category displayed on the homepage.

Change homepage layout to v2. This involved removing the div with class "container" from around the header and main element.
It was on there previously to apply a full-width background the non-footer elements. This is done now by removing the grid-
container class from main and applying selectively to non-full-width elements in the body.