Updates for first round of testing
Changed some generic elements in favour of more semantic elements - header accross all pages is changed to nav, category nav
is a section with ul, and section with class header on patterns page is a header.

Switched page-nav to flexbox because this made it easier to automatically push the 2nd and third elements to the right.

Removed opacity from page nav and sticky header functionality on patterns page as this is changing in v3.