Add future and archived patterns tags
Added future and archived patterns tags on homepage and patterns page.

Also added conditionals in pattern page layout so that it doesn't show blank sections for content that doesn't exist yet.

Changed header on pattern page to flex as well, as it's easier to push content to either side of container element with flexbox.