Change category names and homepage order
Changed category names by adding a 'display_name' value in the categories/_index.md file. This was because:
- Some of the new category display names are quite long - existing code urlizes names for CSS classes, URLs, etc. and using the
new pattern names in the same way would get unwieldy.
- Avoids having to change the category names in all the pattern md files.

There might be a better way to do this, depending on what the final category names are, so I added a TODO in the categories
_index.md file.

This commit also changes the page titles depending on the page.