Changed JS controlling expand/collapse of categories on homepage
Previously, JS would loop through all pattern cards and toggle the display of all non-preview cards. Now, it adds and removes
a class on the category section. The CSS associated with the this class controls the view of non-preview pattern cards.

This avoids the problem I was running into with the Expand All button on the categories view - the fact that everything was
toggled was leading to problems when expand all was clicked.