Move pattern card to a partial
When making changes to pattern card (e.g. moving tags from title to image) previously had to make it in all three homepage
partials.

Moved pattern card code to its own partial so that changes only need to be made in one place.

This involved passing variables to the partial using 'dict' (dictionary), because the category-view uses the range index,
so multiple variables need to be passed. For consistency in naming, used dict on most-recent-view and alphabetical-view as well.