Changed homepage animation for pattern cards to transition opacity instead of container height
Previous animation - changing container height - had performance issues and also required splitting the pattern cards into two
divs - preview and hidden. This caused problems when resizing screens, because the number of cards to be included in the preview
 div was hardcoded at three (fine on desktop and mobile but didn't work on medium screens, where the cards go down to two
across the page).

Changed this so the extra cards are set to display none via CSS, and their display is toggled in JS. Added animation through
changing the opacity slightly later using setTimeout.