Get expand all/collapse all button working
Previous function to control expanding/collapsing categories worked
by toggling the preview class on the section. This didn't work with
the expand all button, because that needed to know what categories
were expanded/collapsed in order to display the right text and selectively
open the unexpanded categories.

Separated expansion and collapsing functionality into two separate
functions and added if statement on see-more-button and expand-all
button click event handlers to determine if category is opened or
closed.

Removed image from testing.