Add example images and related patterns on patterns page
Added space for images (screenshots) to go with pattern examples on
patterns page, if available.

Also added related patterns block on patterns page. This is still WIP
as there are a few bugs to work out (e.g. it can't select itself as a
related pattern, but should still display 2 related patterns).

This commit also removes passing variables to pattern-card partial in
a dict as it was no longer needed with the new way we're showing and
hiding pattern cards.