Add a mobile version of the footer
Created separate files for the mobile and desktop footers, to allow
different order of the content within.

The mobile footer height changes significantly depending on how the text
wraps inside. For this reason I added two additional media queries at
breakpoints which cause extra line wraps. Those points are used to tweak
the heights and margins needed for the footer to have as much space at
it needs.

This change does not affect the grid in any other way.