2nd visual pass - homepage
This commit fixes footer styling and makes small changes to the
homepage. It also adds a new header image.

This commit removes the expand all category button from the category-
nav and includes it only in the category-view partial.

It also takes away the rule that categories with 3 or less patterns
shouldn't show the expand all button - this was causing problems on mobile
because it doesn't collapse to 1 pattern. Removed this for now, maybe to
re-add in future iterations.

Co-authored-by: David Marques <david@projectsbyif.com>