Update Piwik code to point at new server and correct site ID
Piwik code was set for the catalogue when the initial changes were
made to the new site. This was copied directly from the new website
piwik partial and contained the wrong site ID. As a result, visits to
the Catalogue were logged as being made to the main site. In addition,
 an old URL was used for the Piwik code.

Changed site ID to point at the catalogue. Also changed piwik URL to
point at the new server.